Update RenameJPGs.py
This commit reflects the following edits:
(1) Checking tag 0x9003 (DateTime when JPG file was originally created). Only if this fails does it then check tag 0x0132 (DateTime when JPG file was last modified). If that fails, then an error message is issued.
(2) The break statements have been removed from the try block. These were causing the code to exit the for loop prematurely -- i.e., when the user selected multiple JPG files for renaming, only the first selected JPG file was being renamed.
(3) Spelling of "occurred" in the error message was corrected.